+++ GUI/ViewModel/ViewModelUpdateSection.py
